Tera Templating Engine

ExTera is an R6 class object that uses extendr to encapsulate Tera's templating engine. In addition to providing rendering functionality, it acts as a library to hold templates that may include complex dependencies, a feature called template "inheritance" in Tera.

Details

A templating engine requires two things:

a template, as you may have guessed, that includes variables and rendering logic describing where and how to inject data, and

a context, or a set of variables and values to be injected into the template.

Templating syntax is described in the Tera docs.

Methods

Public methods

ExTera$new()

ExTera$print()

ExTera$add_file_templates()

ExTera$add_string_templates()

ExTera$list_templates()

ExTera$render_to_file()

ExTera$render_to_string()

ExTera$autoescape_on()

ExTera$autoescape_off()

ExTera$clone()

Method new()

Create a new ExTera object.

Usage

ExTera$new()

Returns

Self (invisibly)

Method print()

print method for ExTera object.

Usage

ExTera$print(n = 10L, ...)

Arguments

n

integer scalar, number of templates to print (default is 10L)

...

ignored

Returns

Self (invisibly)

Method add_file_templates()

Add templates to library from file paths.

Usage

ExTera$add_file_templates(...)

Arguments

...

specify list of templates as key-value pairs where key is the name of the template and value is the path to the template on file.

Details

All templates must be named.

Returns

Self (invisibly)

Method add_string_templates()

Add templates to library from character strings.

Usage

ExTera$add_string_templates(...)

Arguments

...

specify list of templates as key-value pairs where key is the name of the template and value is a string template.

Details

All templates must be named.

Returns

Self (invisibly)

Method list_templates()

List current templates in library.

Usage

ExTera$list_templates()

Returns

NULL (invisibly)

Method render_to_file()

Render specified template to file.

Usage

ExTera$render_to_file(template, outfile, ...)

Arguments

template

character scalar, the name of the template to render.

outfile

character scalar, the path to file where template is to be rendered.

...

specify context as key-value pairs where key is the template variable and value is the data to inject.

Details

All context elements must be named.

Returns

outfile (invisibly)

Method render_to_string()

Render specified template to string.

Usage

ExTera$render_to_string(template, ...)

Arguments

template

character scalar, the name of the template to render.

...

specify context as key-value pairs where key is the template variable and value is the data to inject.

Details

All context elements must be named.

Returns

Rendered string with class TeraString for "pretty" printing.

Method autoescape_on()

Turn on autoescaping of HTML.

Usage

ExTera$autoescape_on()

Details

Autoescaping is on by default.

Returns

Self (invisibly)

Method autoescape_off()

Turn off autoescaping of HTML.

Usage

ExTera$autoescape_off()

Details

Autoescaping is on by default.

Returns

Self (invisibly)

Method clone()

The objects of this class are cloneable with this method.

Usage

ExTera$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

tera <- ExTera$new()

# from string template
tera$add_string_templates(
  "hello-world" = '<p>Hello {{ x }}. This is {{ y }}.</p>'
)

# to string render
tera$render_to_string(
  "hello-world",
  x = "world",
  y = "ExTera"
)
#> Rendered hello-world template:
#> 
#> <p>Hello world. This is ExTera.</p>
#> 

# from file template
outdir <- tempdir()

tmp <- file.path(
  outdir,
  "hello-world-template.html"
)

cat(
  '<p>Hello {{ x }}. This is {{ y }}.</p>',
  file = tmp
)

tera$add_file_templates(
  "hello-world.html" = tmp
)

outfile <- file.path(
  outdir,
  "hello-world-rendered.html"
)

# to file render
tera$render_to_file(
  "hello-world.html",
  outfile = outfile,
  x = "world",
  y = "ExTera"
)

readLines(outfile)
#> Warning: incomplete final line found on '/tmp/Rtmp52cuud/hello-world-rendered.html'
#> [1] "<p>Hello world. This is ExTera.</p>"